Warnings II: Fix some signed/unsigned mismatches.
Also define the __unused macro for function arguments.

+++ src/json/rangstr.h
@@ -9,7 +9,7 @@
 typedef struct {
 	int type;
 	int next;
-	int f, t;
+	size_t f, t;
 	const char *p;
 } Rangstr;
