+++ src/sdb.h
@@ -61,8 +61,8 @@ void sdb_reset (Sdb *s);
 
 int sdb_query (Sdb *s, const char *cmd);
 int sdb_queryf (Sdb *s, const char *fmt, ...);
-char *sdb_querys (Sdb *s, char *buf, int len, const char *cmd);
-char *sdb_querysf (Sdb *s, char *buf, int buflen, const char *fmt, ...);
+char *sdb_querys (Sdb *s, char *buf, size_t len, const char *cmd);
+char *sdb_querysf (Sdb *s, char *buf, size_t buflen, const char *fmt, ...);
 int sdb_exists (Sdb*, const char *key);
 int sdb_nexists (Sdb*, const char *key);
 int sdb_remove (Sdb*, const char *key, ut32 cas);
@@ -118,8 +118,8 @@ char *sdb_json_unindent(const char *s);
 
 typedef struct {
 	char *buf;
-	int blen;
-	int len;
+	size_t blen;
+	size_t len;
 } SdbJsonString;
 
 const char *sdb_json_format(SdbJsonString* s, const char *fmt, ...);
